fix: resolve test compilation errors in UI test suite
- DataCompletenessServiceTests: use ManifestService.Instance instead of
  new ManifestService() (constructor is private singleton)
- AlertServiceTests: cast AlertSeverity enum to int for
  BeGreaterThanOrEqualTo comparison (FluentAssertions limitation)
- WatchlistServiceTests: replace Instance_CanBeReplaced test with
  Instance_IsAlwaysSameReference (Instance property is read-only)
- StatusEndpointModels: add missing XML documentation comments to
  ProviderLatencyStatsDto and ProviderLatencySummaryDto properties

+++ tests/MarketDataCollector.Ui.Tests/Services/AlertServiceTests.cs
@@ -248,7 +248,7 @@ public sealed class AlertServiceTests
         var groups = svc.GetGroupedAlerts();
         if (groups.Count >= 2)
         {
-            groups[0].Severity.Should().BeGreaterThanOrEqualTo(groups[1].Severity);
+            ((int)groups[0].Severity).Should().BeGreaterThanOrEqualTo((int)groups[1].Severity);
         }
     }
